Added team update history
Team update snapshots are now saved every time a member uploads an update. You can retrieve update history for a team with:

```
/v3/team/{teamNumber}/updates/history
```

This will return a list of all previous updates for the team (not including the current state). The field `modifiedDate` on the response is the date and time that update was originally created.