Goodbye Node.js, hello .net!
For a wide variety of performance reasons, we've decided to migrate this project to .net instead of node. Given that we do not control the upstream endpoints we talk to, we have no control over their performance. We noticed that too often these endpoints would have degraded performance for various reasons, and that would bring the entire node app grinding to a halt with event loop stalls while it was waiting to deal with those responses. We decided instead to switch to a truly multi-threaded language that we have experience with, that being .net.