Adding APIs for FTCScout and Stsatbotics
I've been thinking about the mixed content experience, and if we are to make it work consistently, we need to direct all API traffic through our APIs. This will add three new endpoints, two for FTCScout and one for Statbotics, so that the local client will only need to talk to gatool and the local server (when using FTC local server option or Cheesy Arena). This ensures that if we decide to publish an unsecured path to gatool, these additional services will not be blocked due to mixed content.